refactor: 상품/재고 KafkaConfig 제거 후 통신 정상 확인
