+++ client/product/src/main/java/com/live_commerce/product/inventory/application/service/InventoryService.java
@@ -11,7 +11,7 @@ import com.live_commerce.product.inventory.domain.exception.InventoryException;
 import com.live_commerce.product.inventory.domain.model.Inventory;
 import com.live_commerce.product.inventory.domain.model.InventoryStatus;
 import com.live_commerce.product.inventory.domain.repository.InventoryRepository;
-import com.live_commerce.product.inventory.infrastructure.kafka.event.InventorySoldOutEvent;
+import com.live_commerce.product.product.infrastructure.kafka.event.InventorySoldOutEvent;
 import com.live_commerce.product.product.domain.repository.ProductRepository;
 import com.live_commerce.product.inventory.infrastructure.redisson.DistributedLock;
 import lombok.RequiredArgsConstructor;
