--- client/product/src/main/java/com/live_commerce/product/inventory/infrastructure/config/KafkaConsumerConfig.java
@@ -1,65 +1,0 @@
-package com.live_commerce.product.inventory.infrastructure.config;
-
-import com.live_commerce.product.inventory.infrastructure.kafka.event.OrderCreatedEvent;
-import lombok.RequiredArgsConstructor;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
-import org.springframework.kafka.core.ConsumerFactory;
-import org.springframework.kafka.support.serializer.ErrorHandlingDeserializer;
-import org.springframework.kafka.support.serializer.JsonDeserializer;
-import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
-import org.springframework.kafka.support.serializer.JsonSerializer;
-
-import java.util.HashMap;
-import java.util.Map;
-
-
-@EnableKafka
-@Configuration
-@RequiredArgsConstructor
-public class KafkaConsumerConfig {
-
-    private final KafkaProperties kafkaProperties;
-
-    @Bean
-    public ConsumerFactory<String, Object> inventoryConsumerFactory() {
-
-        JsonDeserializer<Object> deserializer = new JsonDeserializer<>();
-        deserializer.setRemoveTypeHeaders(false);
-        deserializer.addTrustedPackages("*");
-
-        Map<String, Object> configProps = new HashMap<>(kafkaProperties.buildConsumerProperties());
-        configProps.put(ConsumerConfig.GROUP_ID_CONFIG, "inventory-group");
-        return new DefaultKafkaConsumerFactory<>(configProps, new StringDeserializer(), deserializer);
-    }
-
-    @Bean(name = "inventoryKafkaListenerContainerFactory")
-    public ConcurrentKafkaListenerContainerFactory<String, Object> kafkaListenerContainerFactory() {
-        ConcurrentKafkaListenerContainerFactory<String, Object> factory = new ConcurrentKafkaListenerContainerFactory<>();
-        factory.setConsumerFactory(inventoryConsumerFactory());
-        return factory;
-    }
-
-    @Bean
-    public ConsumerFactory<String, Object> productConsumerFactory() {
-        Map<String, Object> configProps = new HashMap<>(kafkaProperties.buildConsumerProperties());
-        configProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        configProps.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
-        configProps.put(ConsumerConfig.GROUP_ID_CONFIG, "product-group");
-        return new DefaultKafkaConsumerFactory<>(configProps);
-    }
-
-    @Bean(name = "productKafkaListenerContainerFactory")
-    public ConcurrentKafkaListenerContainerFactory<String, Object> productKafkaListenerContainerFactory() {
-        ConcurrentKafkaListenerContainerFactory<String, Object> factory = new ConcurrentKafkaListenerContainerFactory<>();
-        factory.setConsumerFactory(productConsumerFactory());
-        return factory;
-    }
-}
