--- client/product/src/main/java/com/live_commerce/product/inventory/infrastructure/kafka/event/OrderCreatedEvent.java
@@ -1,10 +1,0 @@
-package com.live_commerce.product.inventory.infrastructure.kafka.event;
-
-import java.util.UUID;
-
-public record OrderCreatedEvent(
-        UUID orderId,
-        UUID productId,
-        int quantity
-) {
-}
