+++ client/product/src/main/java/com/live_commerce/product/product/infrastructure/kafka/event/OrderRequestedInventoryEvent.java
@@ -1,0 +1,10 @@
+package com.live_commerce.product.product.infrastructure.kafka.event;
+
+import java.util.UUID;
+
+public record OrderRequestedInventoryEvent(
+        UUID orderId,
+        UUID productId,
+        int quantity
+) {
+}
