+++ server/config/src/main/resources/config-repo/product-dev.yml
@@ -33,29 +33,28 @@ spring:
     host: localhost
     port: 6379
 
-    kafka:
-      bootstrap-servers: localhost:9092
-      producer:
-        key-serializer: org.apache.kafka.common.serialization.StringSerializer
-        value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
-        properties:
-          spring.json.type.mapping: >
-            order-created:com.live_commerce.product.inventory.infrastructure.kafka.event.OrderCreatedEvent,
-            inventory-decreased:com.live_commerce.product.inventory.infrastructure.kafka.event.InventoryDecreasedEvent
+  kafka:
+    bootstrap-servers: localhost:9092
+    producer:
+      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
+      properties:
+        spring.json.type.mapping: >
+          inventory-decrease:com.live_commerce.product.product.infrastructure.kafka.event.OrderRequestedInventoryEvent,
+          inventory-decreased:com.live_commerce.product.product.infrastructure.kafka.event.InventoryDecreasedEvent
 
-      consumer:
-        auto-offset-reset: earliest
-        key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-        value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
-        properties:
-          spring.json.trusted.packages: com.live_commerce
-          spring.json.type.mapping: >
-            order-created:com.live_commerce.product.inventory.infrastructure.kafka.event.OrderCreatedEvent,
-            inventory-decreased:com.live_commerce.product.inventory.infrastructure.kafka.event.InventoryDecreasedEvent
-        listener:
-          ack-mode: MANUAL_IMMEDIATE
-        #      reconnect-backoff-ms: 5000
-        #      reconnect-backoff-max-ms: 30000
+    consumer:
+      group-id: ${spring.application.name}
+      auto-offset-reset: earliest
+      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
+      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
+      properties:
+        spring.json.trusted.packages: '*'
+        spring.json.type.mapping: >
+          inventory-decrease:com.live_commerce.product.product.infrastructure.kafka.event.OrderRequestedInventoryEvent,
+          inventory-decreased:com.live_commerce.product.product.infrastructure.kafka.event.InventoryDecreasedEvent
+      listener:
+        ack-mode: MANUAL_IMMEDIATE
 
 gateway:
   base-url: http://localhost:19091
